新規実績、タスクを追加しようとすると発生するバグの修正 アップデート時の不具合修正

+++ src2/UpdateTaskManage.cs
@@ -6,6 +6,7 @@ namespace UpdateTaskManage
     {
         public UpdateTaskManage()
         {
+            InitializeComponent();
             string currentpath = (Application.ExecutablePath).Replace(Path.GetFileName(Application.ExecutablePath), "");
 
             if (!File.Exists(currentpath + "\\TaskManage.zip"))
@@ -14,15 +15,24 @@ namespace UpdateTaskManage
                 this.Close();
             }
 
+            if (Directory.Exists(currentpath + "\\TaskManage"))
+            {
+                Directory.Delete(currentpath + "\\TaskManage", true);
+            }
+
             if (!Directory.Exists(currentpath + "\\TaskManage"))
             {
-                System.IO.Compression.ZipFile.ExtractToDirectory(currentpath + "\\TaskManage.zip", currentpath + "\\TaskManage", System.Text.Encoding.GetEncoding("UTF-8"));
+                System.IO.Compression.ZipFile.ExtractToDirectory(currentpath + "\\TaskManage.zip", currentpath + "\\TaskManage", false);
             }
 
             // ファイル名取得して同じファイルは削除
             string[] files = Directory.GetFiles(currentpath + "\\TaskManage", "*");
             foreach (string file in files)
             {
+                if (file.Contains("UpdateTask"))
+                {
+                    continue;
+                }
                 if (File.Exists(Path.GetFileName(file)))
                 {
                     File.Delete(currentpath + "\\" + Path.GetFileName(file));
@@ -31,12 +41,16 @@ namespace UpdateTaskManage
 
             foreach (string file in files)
             {
+                if (file.Contains("UpdateTask"))
+                {
+                    continue;
+                }
                 File.Move(currentpath + "\\TaskManage\\" + Path.GetFileName(file), currentpath + Path.GetFileName(file));
             }
 
             if (Directory.Exists(currentpath + "\\TaskManage"))
             {
-                Directory.Delete(currentpath + "\\TaskManage");
+                Directory.Delete(currentpath + "\\TaskManage", true);
             }
             Process.Start(currentpath + "\\TaskManage.exe");
             this.Close();
